remove compilation warning
git-svn-id: https://svn.code.sf.net/p/auber/code/branches/tulip_3_0_0_dev@698 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/src/Polynome.cpp
@@ -61,7 +61,7 @@ void Polynome::resolv(float result[3], int &nResult) {
 	
       float alpha = 0;
 	
-      if (abs(k) == 1)
+      if (abs((int) k) == 1)
 	alpha = - PI * (k - 1) / 2;
       else 
 	alpha = acos(k);
